LOGBROKER-10309 Better
