+++ ydb/core/persqueue/public/utils.h
@@ -1,6 +1,5 @@
 #pragma once
 
-#include <deque>
 #include <util/datetime/base.h>
 #include <util/string/builder.h>
 #include <ydb/core/base/appdata_fwd.h>
@@ -88,48 +87,6 @@ TPartitionGraph MakePartitionGraph(const NKikimrSchemeOp::TPersQueueGroupDescrip
 
 TPartitionGraph::TPtr MakeSharedPartitionGraph(const NKikimrPQ::TPQTabletConfig& config);
 TPartitionGraph::TPtr MakeSharedPartitionGraph(const NKikimrSchemeOp::TPersQueueGroupDescription& config);
-
-template <typename T>
-class TLastCounter {
-    static constexpr size_t MaxValueCount = 2;
-
-public:
-    void Use(const T& value, const TInstant& now) {
-        const auto full = MaxValueCount == Values.size();
-        if (!Values.empty() && Values[0].Value == value) {
-            auto& v0 = Values[0];
-            if (v0.LastUseTime < now) {
-                v0.LastUseTime = now;
-                if (full && Values[1].LastUseTime != now) {
-                    Values.push_back(std::move(v0));
-                    Values.pop_front();
-                }
-            }
-        } else if (full && Values[1].Value == value) {
-            Values[1].LastUseTime = now;
-        } else if (!full || Values[0].LastUseTime < now) {
-            if (full) {
-                Values.pop_front();
-            }
-            Values.push_back(Data{now, value});
-        }
-    }
-    size_t Count(const TInstant& expirationTime) {
-        return std::count_if(Values.begin(), Values.end(), [&](const auto& i) {
-            return i.LastUseTime >= expirationTime;
-        });
-    }
-    const TString& LastValue() const {
-        return Values.back().Value;
-    }
-
-private:
-    struct Data {
-        TInstant LastUseTime;
-        T Value;
-    };
-    std::deque<Data> Values;
-};
 
 Y_PURE_FUNCTION bool PreciseReadFromTimestampBehaviourEnabled(const NKikimr::TAppData& appData);
