+++ ydb/core/persqueue/ut/partitioning_keys_manager_ut.cpp
@@ -9,7 +9,7 @@
 #include <util/datetime/base.h>
 #include <util/generic/guid.h>
 #include <util/generic/string.h>
-#include <ydb/core/persqueue/public/partitioning_keys_manager.h>
+#include <ydb/core/persqueue/pqtablet/partition/partitioning_keys_manager.h>
 #include <ydb/services/lib/sharding/sharding.h>
